backend/requirements.txt
fastapi
uvicorn
pillow
transformers
torch
accelerate